tpa/backtrack Relax value type constraints
Because of the requirement for the value type to be convertible to a std::reference_wrapper lvalue unreferenceable values couldn't be returned from the function – for example, a simple int from another function.

We now allow any kind of value type to be used. The user should make the choice of using reference types or values – and should make sure that they live long enough for them to be used in the iteration scope.